Move getFixtures.ts and regenValidFixtures.ts into fixtures/ directory
